+++ discovery/src/asapo_discovery/protocols/protocol_test.go
@@ -25,7 +25,8 @@ var protocolTests = []protocolTest{
 
 
 // producer
-	{"producer", "v0.4", true, "current", "v0.4"},
+	{"producer", "v0.5", true, "current", "v0.5"},
+	{"producer", "v0.4", true, "deprecates", "v0.4"},
 	{"producer", "v0.3", true, "deprecates", "v0.3"},
 	{"producer", "v0.2", true, "deprecates", "v0.2"},
 	{"producer", "v0.1", true, "deprecates", "v0.1"},