+++ receiver/unittests/request_handler/file_processors/test_file_processor.cpp
@@ -11,33 +11,8 @@
 
 #include "../../receiver_mocking.h"
 
-using ::testing::Test;
-using ::testing::Return;
-using ::testing::ReturnRef;
-using ::testing::_;
-using ::testing::DoAll;
-using ::testing::SetArgReferee;
-using ::testing::Gt;
-using ::testing::Eq;
-using ::testing::Ne;
-using ::testing::Mock;
-using ::testing::NiceMock;
-using ::testing::InSequence;
-using ::testing::SetArgPointee;
-using ::testing::AllOf;
-using ::testing::HasSubstr;
-
-
-using ::asapo::Error;
-using ::asapo::GetRootFolder;
-using ::asapo::ErrorInterface;
-using ::asapo::FileDescriptor;
-using ::asapo::SocketDescriptor;
-using ::asapo::MockIO;
-using asapo::Request;
-using asapo::ReceiveFileProcessor;
-using ::asapo::GenericRequestHeader;
-using asapo::MockRequest;
+using namespace testing;
+using namespace asapo;
 
 namespace {
 
@@ -48,8 +23,9 @@ class FileProcessorTests : public Test {
     NiceMock<asapo::MockLogger> mock_logger;
     std::string expected_offline_path =  "offline";
     std::string expected_online_path =  "online";
-    void MockRequestData(std::string fname, asapo::SourceType type);
-    void SetUp() override {
+    void MockRequestData(std::string fname, asapo::SourceType type, bool write_raw_to_core);
+    CustomRequestData expected_custom_data {kDefaultIngestMode, 0, 0};
+   void SetUp() override {
         GenericRequestHeader request_header;
         request_header.data_id = 2;
         asapo::ReceiverConfig test_config;
@@ -61,15 +37,24 @@ class FileProcessorTests : public Test {
 
 };
 
-void FileProcessorTests::MockRequestData(std::string fname, asapo::SourceType type) {
+void FileProcessorTests::MockRequestData(std::string fname, asapo::SourceType type, bool write_raw_to_core) {
 
-    if (type == asapo::SourceType::kProcessed) {
+    if (type == asapo::SourceType::kProcessed || write_raw_to_core) {
         EXPECT_CALL(*mock_request, GetOfflinePath())
         .WillRepeatedly(ReturnRef(expected_offline_path));
     } else {
         EXPECT_CALL(*mock_request, GetOnlinePath())
         .WillRepeatedly(ReturnRef(expected_online_path));
     }
+
+    if (write_raw_to_core) {
+        expected_custom_data[asapo::kPosIngestMode] |= kWriteRawDataToOffline;
+    } else {
+        expected_custom_data[asapo::kPosIngestMode] = kDefaultIngestMode;
+    }
+
+    EXPECT_CALL(*mock_request, GetCustomData_t()).WillRepeatedly(Return(expected_custom_data));
+
 
     EXPECT_CALL(*mock_request, GetSourceType()).WillRepeatedly(Return(type));
 
@@ -88,22 +73,25 @@ TEST_F(FileProcessorTests, RawWriteToRaw) {
 
     struct Test {
         asapo::SourceType type;
+        bool write_raw_to_core;
         std::string filename;
         bool error;
         std::string res;
     };
     std::vector<Test> tests = {
-        Test{asapo::SourceType::kProcessed, repl_sep("processed/bla.text"), false, expected_offline_path},
-        Test{asapo::SourceType::kProcessed, repl_sep("raw/bla.text"), true, ""},
-        Test{asapo::SourceType::kProcessed, repl_sep("processed/../bla.text"), true, ""},
-        Test{asapo::SourceType::kProcessed, repl_sep("bla/bla.text"), true, ""},
-        Test{asapo::SourceType::kProcessed, repl_sep("bla.text"), true, ""},
-        Test{asapo::SourceType::kProcessed, repl_sep("./bla.text"), true, ""},
-        Test{asapo::SourceType::kRaw, repl_sep("raw/bla.text"), false, expected_online_path},
+        Test{asapo::SourceType::kProcessed, false, repl_sep("processed/bla.text"), false, expected_offline_path},
+        Test{asapo::SourceType::kProcessed, true, repl_sep("processed/bla.text"), false, expected_offline_path},
+        Test{asapo::SourceType::kProcessed, false, repl_sep("raw/bla.text"), true, ""},
+        Test{asapo::SourceType::kProcessed, false, repl_sep("processed/../bla.text"), true, ""},
+        Test{asapo::SourceType::kProcessed, false, repl_sep("bla/bla.text"), true, ""},
+        Test{asapo::SourceType::kProcessed, false, repl_sep("bla.text"), true, ""},
+        Test{asapo::SourceType::kProcessed, false, repl_sep("./bla.text"), true, ""},
+        Test{asapo::SourceType::kRaw, false, repl_sep("raw/bla.text"), false, expected_online_path},
+        Test{asapo::SourceType::kRaw, true, repl_sep("raw/bla.text"), false, expected_offline_path},
     };
 
     for (auto& test : tests) {
-        MockRequestData(test.filename, test.type);
+        MockRequestData(test.filename, test.type, test.write_raw_to_core);
         std::string res;
         auto err = GetRootFolder(mock_request.get(), &res);
         if (test.error) {